+++ modules/home/profiles/developer/default.nix
@@ -1,0 +1,7 @@
+{
+  imports = [
+    ../../features/shell
+    ../../features/starship
+    ../../features/aws
+  ];
+}