+++ modules/home/profiles/minimal/default.nix
@@ -1,0 +1,6 @@
+{
+  imports = [
+    ../../features/shell
+    ../../features/starship
+  ];
+}